Split layout for docs by device width
