+++ website/components/docs/Navigation.tsx
@@ -1,21 +1,47 @@
 export const Navigation = () => (
-  <div className="xl:w-5/6 xl:mx-auto">
+  <>
     <ul>
       <li>
-        <h3 className="font-rounded font-bold text-gray-600 uppercase text-lg xl:text-xl">
+        <h3 className="font-rounded font-bold text-gray-600 uppercase text-xl">
           Concept
         </h3>
         <ul>
-          <li className="mt-3">What&apos;s Marp?</li>
-        </ul>
-        <ul>
-          <li className="mt-3">Ecosystem</li>
+          <li className="mt-3">Lorem, ipsum.</li>
+          <li className="mt-3">Quod, harum.</li>
+          <li className="mt-3">Ex, in.</li>
+          <li className="mt-3">Iste, in!</li>
+          <li className="mt-3">Veniam, nemo?</li>
+          <li className="mt-3">Sunt, quasi?</li>
+          <li className="mt-3">Quo, repellat!</li>
+          <li className="mt-3">Facere, quia?</li>
+          <li className="mt-3">Culpa, veniam!</li>
+          <li className="mt-3">Harum, iure.</li>
         </ul>
       </li>
     </ul>
     <ul className="mt-8">
       <li>
-        <h3 className="font-rounded font-bold text-gray-600 uppercase text-lg xl:text-xl">
+        <h3 className="font-rounded font-bold text-gray-600 uppercase text-xl">
+          Concept
+        </h3>
+        <ul>
+          <li className="mt-3">Lorem, ipsum.</li>
+          <li className="mt-3">Quod, harum.</li>
+          <li className="mt-3">Ex, in.</li>
+          <li className="mt-3">Iste, in!</li>
+          <li className="mt-3">Veniam, nemo?</li>
+          <li className="mt-3">Iste, in!</li>
+          <li className="mt-3">Veniam, nemo?</li>
+          <li className="mt-3">Iste, in!</li>
+          <li className="mt-3">Veniam, nemo?</li>
+          <li className="mt-3">Iste, in!</li>
+          <li className="mt-3">Veniam, nemo?</li>
+        </ul>
+      </li>
+    </ul>
+    <ul className="mt-8">
+      <li>
+        <h3 className="font-rounded font-bold text-gray-600 uppercase text-xl">
           Markdown
         </h3>
         <ul>
@@ -23,5 +49,5 @@ export const Navigation = () => (
         </ul>
       </li>
     </ul>
-  </div>
+  </>
 )
